Dockerize RapidPro
This commit adds the necessary files and configuration for running
RapidPro inside docker with linked containers for the Postgres Database
and Redis.  The linking uses env vars in the settings file so that the
docker containers can be deployed and point to other services e.g. AWS
RDS without the need for a separate settings file.

The build and run instructions for docker have been added to the
README.md file